// src/app/shared/model/response/surcharge-response.model.ts
export class SurchargeResponseModel {
  SurchargeID: number;
  CreatedDate: Date;
  Description: string;
  Discount: number;
  ImagesPaths: string;
  IsAll: number;
  Status: number;
  TimeEnd: Date;
  TimeStart: Date;
  Title: string;
  UpdatedDate: Date;
  UserCreated: string;
  UserUpdated: string;

  constructor() {
  }
}